atlas/pyproject.toml
[tool.poetry]
name = "pleiades-atlas"
version = "0.2.0"
description = "Core Infrastructure Library for Pleiades Surveillance Platform"
authors = ["Ahmad Saeed Zaidi"]
readme = "README.md"
license = "MIT"
packages = [{include = "atlas", from = "src"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
aiohttp = "^3.9.3"
orjson = "^3.9.15"

huggingface-hub = {version = "^0.20.0", optional = true}
google-cloud-storage = {version = "^2.14.0", optional = true}
pandas = {version = "^2.2.0", optional = true}
pyarrow = {version = "^15.0.0", optional = true}
prefect = {version = "^2.16.0", optional = true}

[tool.poetry.extras]
hf = ["huggingface-hub", "pandas", "pyarrow"]
gcs = ["google-cloud-storage"]
orchestration = ["prefect"]
all = ["huggingface-hub", "google-cloud-storage", "pandas", "pyarrow", "prefect"]

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true